@@ -13,7 +13,7 @@ class AgentConfig(BaseModel):
     """The name of the llm config to use. If specified, this will override global llm config."""
     classpath: str | None = Field(default=None)
     """The classpath of the agent to use. To be used for custom agents that are not defined in the openhands.agenthub package."""
-    system_prompt_filename: str = Field(default='system_prompt.j2')
+    system_prompt_filename: str = Field(default="system_prompt.j2")
     """Filename of the system prompt template file within the agent's prompt directory. Defaults to 'system_prompt.j2'."""
     enable_browsing: bool = Field(default=True)
     """Whether to enable browsing tool.
@@ -42,12 +42,12 @@ class AgentConfig(BaseModel):
     enable_som_visual_browsing: bool = Field(default=True)
     """Whether to enable SoM (Set of Marks) visual browsing."""
     condenser: CondenserConfig = Field(
-        default_factory=lambda: NoOpCondenserConfig(type='noop')
+        default_factory=lambda: NoOpCondenserConfig(type="noop"),
     )
     extended: ExtendedConfig = Field(default_factory=lambda: ExtendedConfig({}))
     """Extended configuration for the agent."""
 
-    model_config = {'extra': 'forbid'}
+    model_config = {"extra": "forbid"}
 
     @classmethod
     def from_toml_section(cls, data: dict) -> dict[str, AgentConfig]:
@@ -86,29 +86,29 @@ class AgentConfig(BaseModel):
         # Try to create the base config
         try:
             base_config = cls.model_validate(base_data)
-            agent_mapping['agent'] = base_config
+            agent_mapping["agent"] = base_config
         except ValidationError as e:
-            logger.warning(f'Invalid base agent configuration: {e}. Using defaults.')
+            logger.warning(f"Invalid base agent configuration: {e}. Using defaults.")
             # If base config fails, create a default one
             base_config = cls()
             # Still add it to the mapping
-            agent_mapping['agent'] = base_config
+            agent_mapping["agent"] = base_config
 
         # Process each custom section independently
         for name, overrides in custom_sections.items():
             try:
                 # Merge base config with overrides
                 merged = {**base_config.model_dump(), **overrides}
-                if merged.get('classpath'):
+                if merged.get("classpath"):
                     # if an explicit classpath is given, try to load it and look up its config model class
                     from openhands.controller.agent import Agent
 
                     try:
-                        agent_cls = get_impl(Agent, merged.get('classpath'))
+                        agent_cls = get_impl(Agent, merged.get("classpath"))
                         custom_config = agent_cls.config_model.model_validate(merged)
                     except Exception as e:
                         logger.warning(
-                            f'Failed to load custom agent class [{merged.get("classpath")}]: {e}. Using default config model.'
+                            f'Failed to load custom agent class [{merged.get("classpath")}]: {e}. Using default config model.',
                         )
                         custom_config = cls.model_validate(merged)
                 else:
@@ -123,7 +123,7 @@ class AgentConfig(BaseModel):
                 agent_mapping[name] = custom_config
             except ValidationError as e:
                 logger.warning(
-                    f'Invalid agent configuration for [{name}]: {e}. This section will be skipped.'
+                    f"Invalid agent configuration for [{name}]: {e}. This section will be skipped.",
                 )
                 # Skip this custom section but continue with others
                 continue
