@@ -11,31 +11,31 @@ from openhands.core.config.llm_config import LLMConfig
 class NoOpCondenserConfig(BaseModel):
     """Configuration for NoOpCondenser."""
 
-    type: Literal['noop'] = Field('noop')
+    type: Literal["noop"] = Field("noop")
 
-    model_config = {'extra': 'forbid'}
+    model_config = {"extra": "forbid"}
 
 
 class ObservationMaskingCondenserConfig(BaseModel):
     """Configuration for ObservationMaskingCondenser."""
 
-    type: Literal['observation_masking'] = Field('observation_masking')
+    type: Literal["observation_masking"] = Field("observation_masking")
     attention_window: int = Field(
         default=100,
-        description='The number of most-recent events where observations will not be masked.',
+        description="The number of most-recent events where observations will not be masked.",
         ge=1,
     )
 
-    model_config = {'extra': 'forbid'}
+    model_config = {"extra": "forbid"}
 
 
 class BrowserOutputCondenserConfig(BaseModel):
     """Configuration for the BrowserOutputCondenser."""
 
-    type: Literal['browser_output_masking'] = Field('browser_output_masking')
+    type: Literal["browser_output_masking"] = Field("browser_output_masking")
     attention_window: int = Field(
         default=1,
-        description='The number of most recent browser output observations that will not be masked.',
+        description="The number of most recent browser output observations that will not be masked.",
         ge=1,
     )
 
@@ -43,116 +43,116 @@ class BrowserOutputCondenserConfig(BaseModel):
 class RecentEventsCondenserConfig(BaseModel):
     """Configuration for RecentEventsCondenser."""
 
-    type: Literal['recent'] = Field('recent')
+    type: Literal["recent"] = Field("recent")
 
     # at least one event by default, because the best guess is that it is the user task
     keep_first: int = Field(
         default=1,
-        description='The number of initial events to condense.',
+        description="The number of initial events to condense.",
         ge=0,
     )
     max_events: int = Field(
-        default=100, description='Maximum number of events to keep.', ge=1
+        default=100, description="Maximum number of events to keep.", ge=1,
     )
 
-    model_config = {'extra': 'forbid'}
+    model_config = {"extra": "forbid"}
 
 
 class LLMSummarizingCondenserConfig(BaseModel):
     """Configuration for LLMCondenser."""
 
-    type: Literal['llm'] = Field('llm')
+    type: Literal["llm"] = Field("llm")
     llm_config: LLMConfig = Field(
-        ..., description='Configuration for the LLM to use for condensing.'
+        ..., description="Configuration for the LLM to use for condensing.",
     )
 
     # at least one event by default, because the best guess is that it's the user task
     keep_first: int = Field(
         default=1,
-        description='Number of initial events to always keep in history.',
+        description="Number of initial events to always keep in history.",
         ge=0,
     )
     max_size: int = Field(
         default=100,
-        description='Maximum size of the condensed history before triggering forgetting.',
+        description="Maximum size of the condensed history before triggering forgetting.",
         ge=2,
     )
     max_event_length: int = Field(
         default=10_000,
-        description='Maximum length of the event representations to be passed to the LLM.',
+        description="Maximum length of the event representations to be passed to the LLM.",
     )
 
-    model_config = {'extra': 'forbid'}
+    model_config = {"extra": "forbid"}
 
 
 class AmortizedForgettingCondenserConfig(BaseModel):
     """Configuration for AmortizedForgettingCondenser."""
 
-    type: Literal['amortized'] = Field('amortized')
+    type: Literal["amortized"] = Field("amortized")
     max_size: int = Field(
         default=100,
-        description='Maximum size of the condensed history before triggering forgetting.',
+        description="Maximum size of the condensed history before triggering forgetting.",
         ge=2,
     )
 
     # at least one event by default, because the best guess is that it's the user task
     keep_first: int = Field(
         default=1,
-        description='Number of initial events to always keep in history.',
+        description="Number of initial events to always keep in history.",
         ge=0,
     )
 
-    model_config = {'extra': 'forbid'}
+    model_config = {"extra": "forbid"}
 
 
 class LLMAttentionCondenserConfig(BaseModel):
     """Configuration for LLMAttentionCondenser."""
 
-    type: Literal['llm_attention'] = Field('llm_attention')
+    type: Literal["llm_attention"] = Field("llm_attention")
     llm_config: LLMConfig = Field(
-        ..., description='Configuration for the LLM to use for attention.'
+        ..., description="Configuration for the LLM to use for attention.",
     )
     max_size: int = Field(
         default=100,
-        description='Maximum size of the condensed history before triggering forgetting.',
+        description="Maximum size of the condensed history before triggering forgetting.",
         ge=2,
     )
 
     # at least one event by default, because the best guess is that it's the user task
     keep_first: int = Field(
         default=1,
-        description='Number of initial events to always keep in history.',
+        description="Number of initial events to always keep in history.",
         ge=0,
     )
 
-    model_config = {'extra': 'forbid'}
+    model_config = {"extra": "forbid"}
 
 
 class StructuredSummaryCondenserConfig(BaseModel):
     """Configuration for StructuredSummaryCondenser instances."""
 
-    type: Literal['structured'] = Field('structured')
+    type: Literal["structured"] = Field("structured")
     llm_config: LLMConfig = Field(
-        ..., description='Configuration for the LLM to use for condensing.'
+        ..., description="Configuration for the LLM to use for condensing.",
     )
 
     # at least one event by default, because the best guess is that it's the user task
     keep_first: int = Field(
         default=1,
-        description='Number of initial events to always keep in history.',
+        description="Number of initial events to always keep in history.",
         ge=0,
     )
     max_size: int = Field(
         default=100,
-        description='Maximum size of the condensed history before triggering forgetting.',
+        description="Maximum size of the condensed history before triggering forgetting.",
         ge=2,
     )
     max_event_length: int = Field(
         default=10_000,
-        description='Maximum length of the event representations to be passed to the LLM.',
+        description="Maximum length of the event representations to be passed to the LLM.",
     )
 
-    model_config = {'extra': 'forbid'}
+    model_config = {"extra": "forbid"}
 
 
 class CondenserPipelineConfig(BaseModel):
@@ -161,13 +161,13 @@ class CondenserPipelineConfig(BaseModel):
     Not currently supported by the TOML or ENV_VAR configuration strategies.
     """
 
-    type: Literal['pipeline'] = Field('pipeline')
+    type: Literal["pipeline"] = Field("pipeline")
     condensers: list[CondenserConfig] = Field(
         default_factory=list,
-        description='List of condenser configurations to be used in the pipeline.',
+        description="List of condenser configurations to be used in the pipeline.",
     )
 
-    model_config = {'extra': 'forbid'}
+    model_config = {"extra": "forbid"}
 
 
 # Type alias for convenience
@@ -185,7 +185,7 @@ CondenserConfig = (
 
 
 def condenser_config_from_toml_section(
-    data: dict, llm_configs: dict | None = None
+    data: dict, llm_configs: dict | None = None,
 ) -> dict[str, CondenserConfig]:
     """
     Create a CondenserConfig instance from a toml dictionary representing the [condenser] section.
@@ -217,41 +217,41 @@ def condenser_config_from_toml_section(
     # Process config
     try:
         # Determine which condenser type to use based on 'type' field
-        condenser_type = data.get('type', 'noop')
+        condenser_type = data.get("type", "noop")
 
         # Handle LLM config reference if needed
         if (
-            condenser_type in ('llm', 'llm_attention')
-            and 'llm_config' in data
-            and isinstance(data['llm_config'], str)
+            condenser_type in ("llm", "llm_attention")
+            and "llm_config" in data
+            and isinstance(data["llm_config"], str)
         ):
-            llm_config_name = data['llm_config']
+            llm_config_name = data["llm_config"]
             if llm_configs and llm_config_name in llm_configs:
                 # Replace the string reference with the actual LLMConfig object
                 data_copy = data.copy()
-                data_copy['llm_config'] = llm_configs[llm_config_name]
+                data_copy["llm_config"] = llm_configs[llm_config_name]
                 config = create_condenser_config(condenser_type, data_copy)
             else:
                 logger.openhands_logger.warning(
-                    f"LLM config '{llm_config_name}' not found for condenser. Using default LLMConfig."
+                    f"LLM config '{llm_config_name}' not found for condenser. Using default LLMConfig.",
                 )
                 # Create a default LLMConfig if the referenced one doesn't exist
                 data_copy = data.copy()
                 # Try to use the fallback 'llm' config
                 if llm_configs is not None:
-                    data_copy['llm_config'] = llm_configs.get('llm')
+                    data_copy["llm_config"] = llm_configs.get("llm")
                 config = create_condenser_config(condenser_type, data_copy)
         else:
             config = create_condenser_config(condenser_type, data)
 
-        condenser_mapping['condenser'] = config
+        condenser_mapping["condenser"] = config
     except (ValidationError, ValueError) as e:
         logger.openhands_logger.warning(
-            f'Invalid condenser configuration: {e}. Using NoOpCondenserConfig.'
+            f"Invalid condenser configuration: {e}. Using NoOpCondenserConfig.",
         )
         # Default to NoOpCondenserConfig if config fails
-        config = NoOpCondenserConfig(type='noop')
-        condenser_mapping['condenser'] = config
+        config = NoOpCondenserConfig(type="noop")
+        condenser_mapping["condenser"] = config
 
     return condenser_mapping
 
@@ -277,17 +277,17 @@ def create_condenser_config(condenser_type: str, data: dict) -> CondenserConfig:
     """
     # Mapping of condenser types to their config classes
     condenser_classes = {
-        'noop': NoOpCondenserConfig,
-        'observation_masking': ObservationMaskingCondenserConfig,
-        'recent': RecentEventsCondenserConfig,
-        'llm': LLMSummarizingCondenserConfig,
-        'amortized': AmortizedForgettingCondenserConfig,
-        'llm_attention': LLMAttentionCondenserConfig,
-        'structured': StructuredSummaryCondenserConfig,
+        "noop": NoOpCondenserConfig,
+        "observation_masking": ObservationMaskingCondenserConfig,
+        "recent": RecentEventsCondenserConfig,
+        "llm": LLMSummarizingCondenserConfig,
+        "amortized": AmortizedForgettingCondenserConfig,
+        "llm_attention": LLMAttentionCondenserConfig,
+        "structured": StructuredSummaryCondenserConfig,
     }
 
     if condenser_type not in condenser_classes:
-        raise ValueError(f'Unknown condenser type: {condenser_type}')
+        raise ValueError(f"Unknown condenser type: {condenser_type}")
 
     # Create and validate the config using direct instantiation
     # Explicitly handle ValidationError to provide more context
@@ -299,5 +299,5 @@ def create_condenser_config(condenser_type: str, data: dict) -> CondenserConfig:
         # Just re-raise with a more descriptive message, but don't try to pass the errors
         # which can cause compatibility issues with different pydantic versions
         raise ValueError(
-            f"Validation failed for condenser type '{condenser_type}': {e}"
+            f"Validation failed for condenser type '{condenser_type}': {e}",
         )
