@@ -4,7 +4,7 @@ from typing import Any, get_args, get_origin
 from pydantic import BaseModel
 from pydantic.fields import FieldInfo
 
-OH_DEFAULT_AGENT = 'CodeActAgent'
+OH_DEFAULT_AGENT = "CodeActAgent"
 OH_MAX_ITERATIONS = 500
 
 
@@ -26,7 +26,7 @@ def get_field_info(field: FieldInfo) -> dict[str, Any]:
     if get_origin(field_type) is UnionType:
         types = get_args(field_type)
         non_none_arg = next(
-            (t for t in types if t is not None and t is not type(None)), None
+            (t for t in types if t is not None and t is not type(None)), None,
         )
         if non_none_arg is not None:
             field_type = non_none_arg
@@ -37,7 +37,7 @@ def get_field_info(field: FieldInfo) -> dict[str, Any]:
         str(field_type)
         if field_type is None
         else (
-            field_type.__name__ if hasattr(field_type, '__name__') else str(field_type)
+            field_type.__name__ if hasattr(field_type, "__name__") else str(field_type)
         )
     )
 
@@ -45,7 +45,7 @@ def get_field_info(field: FieldInfo) -> dict[str, Any]:
     default = field.default
 
     # return a schema with the useful info for frontend
-    return {'type': type_name.lower(), 'optional': optional, 'default': default}
+    return {"type": type_name.lower(), "optional": optional, "default": default}
 
 
 def model_defaults_to_dict(model: BaseModel) -> dict[str, Any]:
