@@ -12,10 +12,10 @@ class SecurityConfig(BaseModel):
     confirmation_mode: bool = Field(default=False)
     security_analyzer: str | None = Field(default=None)
 
-    model_config = {'extra': 'forbid'}
+    model_config = {"extra": "forbid"}
 
     @classmethod
-    def from_toml_section(cls, data: dict) -> dict[str, 'SecurityConfig']:
+    def from_toml_section(cls, data: dict) -> dict[str, "SecurityConfig"]:
         """
         Create a mapping of SecurityConfig instances from a toml dictionary representing the [security] section.
 
@@ -30,8 +30,8 @@ class SecurityConfig(BaseModel):
 
         # Try to create the configuration instance
         try:
-            security_mapping['security'] = cls.model_validate(data)
+            security_mapping["security"] = cls.model_validate(data)
         except ValidationError as e:
-            raise ValueError(f'Invalid security configuration: {e}')
+            raise ValueError(f"Invalid security configuration: {e}")
 
         return security_mapping
