@@ -4,8 +4,9 @@ import pathlib
 import platform
 import sys
 from ast import literal_eval
+from collections.abc import MutableMapping
 from types import UnionType
-from typing import MutableMapping, get_args, get_origin
+from typing import get_args, get_origin
 from uuid import uuid4
 
 import toml
@@ -35,12 +36,12 @@ from openhands.storage import get_file_store
 from openhands.storage.files import FileStore
 from openhands.utils.import_utils import get_impl
 
-JWT_SECRET = '.jwt_secret'
+JWT_SECRET = ".jwt_secret"
 load_dotenv()
 
 
 def load_from_env(
-    cfg: OpenHandsConfig, env_or_toml_dict: dict | MutableMapping[str, str]
+    cfg: OpenHandsConfig, env_or_toml_dict: dict | MutableMapping[str, str],
 ) -> None:
     """Sets config attributes from environment variables or TOML dictionary.
 
@@ -65,7 +66,7 @@ def load_from_env(
         return None
 
     # helper function to set attributes based on env vars
-    def set_attr_from_env(sub_config: BaseModel, prefix: str = '') -> None:
+    def set_attr_from_env(sub_config: BaseModel, prefix: str = "") -> None:
         """Set attributes of a config model based on environment variables."""
         for field_name, field_info in sub_config.model_fields.items():
             field_value = getattr(sub_config, field_name)
@@ -76,7 +77,7 @@ def load_from_env(
             env_var_name = (prefix + field_name).upper()
 
             if isinstance(field_value, BaseModel):
-                set_attr_from_env(field_value, prefix=field_name + '_')
+                set_attr_from_env(field_value, prefix=field_name + "_")
 
             elif env_var_name in env_or_toml_dict:
                 # convert the env var to the correct type and set it
@@ -93,7 +94,7 @@ def load_from_env(
 
                     # Attempt to cast the env var to type hinted in the dataclass
                     if field_type is bool:
-                        cast_value = str(value).lower() in ['true', '1']
+                        cast_value = str(value).lower() in ["true", "1"]
                     # parse dicts and lists like SANDBOX_RUNTIME_STARTUP_ENV_VARS and SANDBOX_RUNTIME_EXTRA_BUILD_ARGS                                                                                                                                     │
                     elif (
                         get_origin(field_type) is dict
@@ -102,13 +103,12 @@ def load_from_env(
                         or field_type is list
                     ):
                         cast_value = literal_eval(value)
-                    else:
-                        if field_type is not None:
-                            cast_value = field_type(value)
+                    elif field_type is not None:
+                        cast_value = field_type(value)
                     setattr(sub_config, field_name, cast_value)
                 except (ValueError, TypeError):
                     logger.openhands_logger.error(
-                        f'Error setting env var {env_var_name}={value}: check that the value is of the right type'
+                        f"Error setting env var {env_var_name}={value}: check that the value is of the right type",
                     )
 
     # Start processing from the root of the config object
@@ -116,13 +116,13 @@ def load_from_env(
 
     # load default LLM config from env
     default_llm_config = cfg.get_llm_config()
-    set_attr_from_env(default_llm_config, 'LLM_')
+    set_attr_from_env(default_llm_config, "LLM_")
     # load default agent config from env
     default_agent_config = cfg.get_agent_config()
-    set_attr_from_env(default_agent_config, 'AGENT_')
+    set_attr_from_env(default_agent_config, "AGENT_")
 
 
-def load_from_toml(cfg: OpenHandsConfig, toml_file: str = 'config.toml') -> None:
+def load_from_toml(cfg: OpenHandsConfig, toml_file: str = "config.toml") -> None:
     """Load the config from the toml file. Supports both styles of config vars.
 
     Args:
@@ -134,24 +134,24 @@ def load_from_toml(cfg: OpenHandsConfig, toml_file: str = 'config.toml') -> None
     """
     # try to read the config.toml file into the config object
     try:
-        with open(toml_file, 'r', encoding='utf-8') as toml_contents:
+        with open(toml_file, encoding="utf-8") as toml_contents:
             toml_config = toml.load(toml_contents)
     except FileNotFoundError:
         return
     except toml.TomlDecodeError as e:
         logger.openhands_logger.warning(
-            f'Cannot parse config from toml, toml values have not been applied.\nError: {e}',
+            f"Cannot parse config from toml, toml values have not been applied.\nError: {e}",
         )
         return
 
     # Check for the [core] section
-    if 'core' not in toml_config:
+    if "core" not in toml_config:
         logger.openhands_logger.warning(
-            f'No [core] section found in {toml_file}. Core settings will use defaults.'
+            f"No [core] section found in {toml_file}. Core settings will use defaults.",
         )
         core_config = {}
     else:
-        core_config = toml_config['core']
+        core_config = toml_config["core"]
 
     # Process core section if present
     for key, value in core_config.items():
@@ -159,107 +159,107 @@ def load_from_toml(cfg: OpenHandsConfig, toml_file: str = 'config.toml') -> None
             setattr(cfg, key, value)
         else:
             logger.openhands_logger.warning(
-                f'Unknown config key "{key}" in [core] section'
+                f'Unknown config key "{key}" in [core] section',
             )
 
     # Process agent section if present
-    if 'agent' in toml_config:
+    if "agent" in toml_config:
         try:
-            agent_mapping = AgentConfig.from_toml_section(toml_config['agent'])
+            agent_mapping = AgentConfig.from_toml_section(toml_config["agent"])
             for agent_key, agent_conf in agent_mapping.items():
                 cfg.set_agent_config(agent_conf, agent_key)
         except (TypeError, KeyError, ValidationError) as e:
             logger.openhands_logger.warning(
-                f'Cannot parse [agent] config from toml, values have not been applied.\nError: {e}'
+                f"Cannot parse [agent] config from toml, values have not been applied.\nError: {e}",
             )
 
     # Process llm section if present
-    if 'llm' in toml_config:
+    if "llm" in toml_config:
         try:
-            llm_mapping = LLMConfig.from_toml_section(toml_config['llm'])
+            llm_mapping = LLMConfig.from_toml_section(toml_config["llm"])
             for llm_key, llm_conf in llm_mapping.items():
                 cfg.set_llm_config(llm_conf, llm_key)
         except (TypeError, KeyError, ValidationError) as e:
             logger.openhands_logger.warning(
-                f'Cannot parse [llm] config from toml, values have not been applied.\nError: {e}'
+                f"Cannot parse [llm] config from toml, values have not been applied.\nError: {e}",
             )
 
     # Process security section if present
-    if 'security' in toml_config:
+    if "security" in toml_config:
         try:
-            security_mapping = SecurityConfig.from_toml_section(toml_config['security'])
+            security_mapping = SecurityConfig.from_toml_section(toml_config["security"])
             # We only use the base security config for now
-            if 'security' in security_mapping:
-                cfg.security = security_mapping['security']
+            if "security" in security_mapping:
+                cfg.security = security_mapping["security"]
         except (TypeError, KeyError, ValidationError) as e:
             logger.openhands_logger.warning(
-                f'Cannot parse [security] config from toml, values have not been applied.\nError: {e}'
+                f"Cannot parse [security] config from toml, values have not been applied.\nError: {e}",
             )
         except ValueError:
             # Re-raise ValueError from SecurityConfig.from_toml_section
-            raise ValueError('Error in [security] section in config.toml')
+            raise ValueError("Error in [security] section in config.toml")
 
     # Process sandbox section if present
-    if 'sandbox' in toml_config:
+    if "sandbox" in toml_config:
         try:
-            sandbox_mapping = SandboxConfig.from_toml_section(toml_config['sandbox'])
+            sandbox_mapping = SandboxConfig.from_toml_section(toml_config["sandbox"])
             # We only use the base sandbox config for now
-            if 'sandbox' in sandbox_mapping:
-                cfg.sandbox = sandbox_mapping['sandbox']
+            if "sandbox" in sandbox_mapping:
+                cfg.sandbox = sandbox_mapping["sandbox"]
         except (TypeError, KeyError, ValidationError) as e:
             logger.openhands_logger.warning(
-                f'Cannot parse [sandbox] config from toml, values have not been applied.\nError: {e}'
+                f"Cannot parse [sandbox] config from toml, values have not been applied.\nError: {e}",
             )
         except ValueError:
             # Re-raise ValueError from SandboxConfig.from_toml_section
-            raise ValueError('Error in [sandbox] section in config.toml')
+            raise ValueError("Error in [sandbox] section in config.toml")
 
     # Process MCP sections if present
-    if 'mcp' in toml_config:
+    if "mcp" in toml_config:
         try:
-            mcp_mapping = MCPConfig.from_toml_section(toml_config['mcp'])
+            mcp_mapping = MCPConfig.from_toml_section(toml_config["mcp"])
             # We only use the base mcp config for now
-            if 'mcp' in mcp_mapping:
-                cfg.mcp = mcp_mapping['mcp']
+            if "mcp" in mcp_mapping:
+                cfg.mcp = mcp_mapping["mcp"]
         except (TypeError, KeyError, ValidationError) as e:
             logger.openhands_logger.warning(
-                f'Cannot parse MCP config from toml, values have not been applied.\nError: {e}'
+                f"Cannot parse MCP config from toml, values have not been applied.\nError: {e}",
             )
         except ValueError:
             # Re-raise ValueError from MCPConfig.from_toml_section
-            raise ValueError('Error in MCP sections in config.toml')
+            raise ValueError("Error in MCP sections in config.toml")
 
     # Process kubernetes section if present
-    if 'kubernetes' in toml_config:
+    if "kubernetes" in toml_config:
         try:
             kubernetes_mapping = KubernetesConfig.from_toml_section(
-                toml_config['kubernetes']
+                toml_config["kubernetes"],
             )
-            if 'kubernetes' in kubernetes_mapping:
-                cfg.kubernetes = kubernetes_mapping['kubernetes']
+            if "kubernetes" in kubernetes_mapping:
+                cfg.kubernetes = kubernetes_mapping["kubernetes"]
         except (TypeError, KeyError, ValidationError) as e:
             logger.openhands_logger.warning(
-                f'Cannot parse [kubernetes] config from toml, values have not been applied.\nError: {e}'
+                f"Cannot parse [kubernetes] config from toml, values have not been applied.\nError: {e}",
             )
 
     # Process condenser section if present
-    if 'condenser' in toml_config:
+    if "condenser" in toml_config:
         try:
             # Pass the LLM configs to the condenser config parser
             condenser_mapping = condenser_config_from_toml_section(
-                toml_config['condenser'], cfg.llms
+                toml_config["condenser"], cfg.llms,
             )
             # Assign the default condenser configuration to the default agent configuration
-            if 'condenser' in condenser_mapping:
+            if "condenser" in condenser_mapping:
                 # Get the default agent config and assign the condenser config to it
                 default_agent_config = cfg.get_agent_config()
-                default_agent_config.condenser = condenser_mapping['condenser']
+                default_agent_config.condenser = condenser_mapping["condenser"]
                 logger.openhands_logger.debug(
-                    'Default condenser configuration loaded from config toml and assigned to default agent'
+                    "Default condenser configuration loaded from config toml and assigned to default agent",
                 )
         except (TypeError, KeyError, ValidationError) as e:
             logger.openhands_logger.warning(
-                f'Cannot parse [condenser] config from toml, values have not been applied.\nError: {e}'
+                f"Cannot parse [condenser] config from toml, values have not been applied.\nError: {e}",
             )
     # If no condenser section is in toml but enable_default_condenser is True,
     # set LLMSummarizingCondenserConfig as default
@@ -272,39 +272,39 @@ def load_from_toml(cfg: OpenHandsConfig, toml_file: str = 'config.toml') -> None
         # Create default LLM summarizing condenser config
         default_condenser = LLMSummarizingCondenserConfig(
             llm_config=cfg.get_llm_config(),  # Use default LLM config
-            type='llm',
+            type="llm",
         )
 
         # Set as default condenser
         default_agent_config.condenser = default_condenser
         logger.openhands_logger.debug(
-            'Default LLM summarizing condenser assigned to default agent (no condenser in config)'
+            "Default LLM summarizing condenser assigned to default agent (no condenser in config)",
         )
 
     # Process extended section if present
-    if 'extended' in toml_config:
+    if "extended" in toml_config:
         try:
-            cfg.extended = ExtendedConfig(toml_config['extended'])
+            cfg.extended = ExtendedConfig(toml_config["extended"])
         except (TypeError, KeyError, ValidationError) as e:
             logger.openhands_logger.warning(
-                f'Cannot parse [extended] config from toml, values have not been applied.\nError: {e}'
+                f"Cannot parse [extended] config from toml, values have not been applied.\nError: {e}",
             )
 
     # Check for unknown sections
     known_sections = {
-        'core',
-        'extended',
-        'agent',
-        'llm',
-        'security',
-        'sandbox',
-        'condenser',
-        'mcp',
-        'kubernetes',
+        "core",
+        "extended",
+        "agent",
+        "llm",
+        "security",
+        "sandbox",
+        "condenser",
+        "mcp",
+        "kubernetes",
     }
     for key in toml_config:
         if key.lower() not in known_sections:
-            logger.openhands_logger.warning(f'Unknown section [{key}] in {toml_file}')
+            logger.openhands_logger.warning(f"Unknown section [{key}] in {toml_file}")
 
 
 def get_or_create_jwt_secret(file_store: FileStore) -> str:
@@ -322,24 +322,24 @@ def finalize_config(cfg: OpenHandsConfig) -> None:
     # Handle the sandbox.volumes parameter
     if cfg.workspace_base is not None or cfg.workspace_mount_path is not None:
         logger.openhands_logger.warning(
-            'DEPRECATED: The WORKSPACE_BASE and WORKSPACE_MOUNT_PATH environment variables are deprecated. '
-            "Please use RUNTIME_MOUNT instead, e.g. 'RUNTIME_MOUNT=/my/host/dir:/workspace:rw'"
+            "DEPRECATED: The WORKSPACE_BASE and WORKSPACE_MOUNT_PATH environment variables are deprecated. "
+            "Please use RUNTIME_MOUNT instead, e.g. 'RUNTIME_MOUNT=/my/host/dir:/workspace:rw'",
         )
     if cfg.sandbox.volumes is not None:
         # Split by commas to handle multiple mounts
-        mounts = cfg.sandbox.volumes.split(',')
+        mounts = cfg.sandbox.volumes.split(",")
 
         # Check if any mount explicitly targets /workspace
         workspace_mount_found = False
         for mount in mounts:
-            parts = mount.split(':')
-            if len(parts) >= 2 and parts[1] == '/workspace':
+            parts = mount.split(":")
+            if len(parts) >= 2 and parts[1] == "/workspace":
                 workspace_mount_found = True
                 host_path = os.path.abspath(parts[0])
 
                 # Set the workspace_mount_path and workspace_mount_path_in_sandbox
                 cfg.workspace_mount_path = host_path
-                cfg.workspace_mount_path_in_sandbox = '/workspace'
+                cfg.workspace_mount_path_in_sandbox = "/workspace"
 
                 # Also set workspace_base
                 cfg.workspace_base = host_path
@@ -349,8 +349,8 @@ def finalize_config(cfg: OpenHandsConfig) -> None:
         # This allows users to mount volumes without affecting the workspace
         if not workspace_mount_found:
             logger.openhands_logger.debug(
-                'No explicit /workspace mount found in SANDBOX_VOLUMES. '
-                'Using default workspace path in sandbox.'
+                "No explicit /workspace mount found in SANDBOX_VOLUMES. "
+                "Using default workspace path in sandbox.",
             )
             # Ensure workspace_mount_path and workspace_base are None to avoid
             # unintended mounting behavior
@@ -359,11 +359,11 @@ def finalize_config(cfg: OpenHandsConfig) -> None:
 
         # Validate all mounts
         for mount in mounts:
-            parts = mount.split(':')
+            parts = mount.split(":")
             if len(parts) < 2 or len(parts) > 3:
                 raise ValueError(
-                    f'Invalid mount format in sandbox.volumes: {mount}. '
-                    f"Expected format: 'host_path:container_path[:mode]', e.g. '/my/host/dir:/workspace:rw'"
+                    f"Invalid mount format in sandbox.volumes: {mount}. "
+                    f"Expected format: 'host_path:container_path[:mode]', e.g. '/my/host/dir:/workspace:rw'",
                 )
 
     # Handle the deprecated workspace_* parameters
@@ -375,17 +375,17 @@ def finalize_config(cfg: OpenHandsConfig) -> None:
 
         if cfg.workspace_mount_rewrite:
             base = cfg.workspace_base or os.getcwd()
-            parts = cfg.workspace_mount_rewrite.split(':')
+            parts = cfg.workspace_mount_rewrite.split(":")
             cfg.workspace_mount_path = base.replace(parts[0], parts[1])
 
     # make sure log_completions_folder is an absolute path
     for llm in cfg.llms.values():
         llm.log_completions_folder = os.path.abspath(llm.log_completions_folder)
 
-    if cfg.sandbox.use_host_network and platform.system() == 'Darwin':
+    if cfg.sandbox.use_host_network and platform.system() == "Darwin":
         logger.openhands_logger.warning(
-            'Please upgrade to Docker Desktop 4.29.0 or later to use host network mode on macOS. '
-            'See https://github.com/docker/roadmap/issues/238#issuecomment-2044688144 for more information.'
+            "Please upgrade to Docker Desktop 4.29.0 or later to use host network mode on macOS. "
+            "See https://github.com/docker/roadmap/issues/238#issuecomment-2044688144 for more information.",
         )
 
     # make sure cache dir exists
@@ -395,26 +395,26 @@ def finalize_config(cfg: OpenHandsConfig) -> None:
     if not cfg.jwt_secret:
         cfg.jwt_secret = SecretStr(
             get_or_create_jwt_secret(
-                get_file_store(cfg.file_store, cfg.file_store_path)
-            )
+                get_file_store(cfg.file_store, cfg.file_store_path),
+            ),
         )
 
     # If CLIRuntime is selected, disable Jupyter for all agents
     # Assuming 'cli' is the identifier for CLIRuntime
-    if cfg.runtime and cfg.runtime.lower() == 'cli':
+    if cfg.runtime and cfg.runtime.lower() == "cli":
         for age_nt_name, agent_config in cfg.agents.items():
             if agent_config.enable_jupyter:
                 agent_config.enable_jupyter = False
             if agent_config.enable_browsing:
                 agent_config.enable_browsing = False
         logger.openhands_logger.debug(
-            'Automatically disabled Jupyter plugin and browsing for all agents '
-            'because CLIRuntime is selected and does not support IPython execution.'
+            "Automatically disabled Jupyter plugin and browsing for all agents "
+            "because CLIRuntime is selected and does not support IPython execution.",
         )
 
 
 def get_agent_config_arg(
-    agent_config_arg: str, toml_file: str = 'config.toml'
+    agent_config_arg: str, toml_file: str = "config.toml",
 ) -> AgentConfig | None:
     """Get a group of agent settings from the config file.
 
@@ -438,36 +438,36 @@ def get_agent_config_arg(
         AgentConfig: The AgentConfig object with the settings from the config file.
     """
     # keep only the name, just in case
-    agent_config_arg = agent_config_arg.strip('[]')
+    agent_config_arg = agent_config_arg.strip("[]")
 
     # truncate the prefix, just in case
-    if agent_config_arg.startswith('agent.'):
+    if agent_config_arg.startswith("agent."):
         agent_config_arg = agent_config_arg[6:]
 
-    logger.openhands_logger.debug(f'Loading agent config from {agent_config_arg}')
+    logger.openhands_logger.debug(f"Loading agent config from {agent_config_arg}")
 
     # load the toml file
     try:
-        with open(toml_file, 'r', encoding='utf-8') as toml_contents:
+        with open(toml_file, encoding="utf-8") as toml_contents:
             toml_config = toml.load(toml_contents)
     except FileNotFoundError as e:
-        logger.openhands_logger.error(f'Config file not found: {e}')
+        logger.openhands_logger.error(f"Config file not found: {e}")
         return None
     except toml.TomlDecodeError as e:
         logger.openhands_logger.error(
-            f'Cannot parse agent group from {agent_config_arg}. Exception: {e}'
+            f"Cannot parse agent group from {agent_config_arg}. Exception: {e}",
         )
         return None
 
     # update the agent config with the specified section
-    if 'agent' in toml_config and agent_config_arg in toml_config['agent']:
-        return AgentConfig(**toml_config['agent'][agent_config_arg])
-    logger.openhands_logger.debug(f'Loading from toml failed for {agent_config_arg}')
+    if "agent" in toml_config and agent_config_arg in toml_config["agent"]:
+        return AgentConfig(**toml_config["agent"][agent_config_arg])
+    logger.openhands_logger.debug(f"Loading from toml failed for {agent_config_arg}")
     return None
 
 
 def get_llm_config_arg(
-    llm_config_arg: str, toml_file: str = 'config.toml'
+    llm_config_arg: str, toml_file: str = "config.toml",
 ) -> LLMConfig | None:
     """Get a group of llm settings from the config file.
 
@@ -495,36 +495,36 @@ def get_llm_config_arg(
         LLMConfig: The LLMConfig object with the settings from the config file.
     """
     # keep only the name, just in case
-    llm_config_arg = llm_config_arg.strip('[]')
+    llm_config_arg = llm_config_arg.strip("[]")
 
     # truncate the prefix, just in case
-    if llm_config_arg.startswith('llm.'):
+    if llm_config_arg.startswith("llm."):
         llm_config_arg = llm_config_arg[4:]
 
-    logger.openhands_logger.debug(f'Loading llm config from {llm_config_arg}')
+    logger.openhands_logger.debug(f"Loading llm config from {llm_config_arg}")
 
     # load the toml file
     try:
-        with open(toml_file, 'r', encoding='utf-8') as toml_contents:
+        with open(toml_file, encoding="utf-8") as toml_contents:
             toml_config = toml.load(toml_contents)
     except FileNotFoundError as e:
-        logger.openhands_logger.error(f'Config file not found: {e}')
+        logger.openhands_logger.error(f"Config file not found: {e}")
         return None
     except toml.TomlDecodeError as e:
         logger.openhands_logger.error(
-            f'Cannot parse llm group from {llm_config_arg}. Exception: {e}'
+            f"Cannot parse llm group from {llm_config_arg}. Exception: {e}",
         )
         return None
 
     # update the llm config with the specified section
-    if 'llm' in toml_config and llm_config_arg in toml_config['llm']:
-        return LLMConfig(**toml_config['llm'][llm_config_arg])
-    logger.openhands_logger.debug(f'Loading from toml failed for {llm_config_arg}')
+    if "llm" in toml_config and llm_config_arg in toml_config["llm"]:
+        return LLMConfig(**toml_config["llm"][llm_config_arg])
+    logger.openhands_logger.debug(f"Loading from toml failed for {llm_config_arg}")
     return None
 
 
 def get_condenser_config_arg(
-    condenser_config_arg: str, toml_file: str = 'config.toml'
+    condenser_config_arg: str, toml_file: str = "config.toml",
 ) -> CondenserConfig | None:
     """Get a group of condenser settings from the config file by name.
 
@@ -553,71 +553,71 @@ def get_condenser_config_arg(
         CondenserConfig: The CondenserConfig object with the settings from the config file, or None if not found/error.
     """
     # keep only the name, just in case
-    condenser_config_arg = condenser_config_arg.strip('[]')
+    condenser_config_arg = condenser_config_arg.strip("[]")
 
     # truncate the prefix, just in case
-    if condenser_config_arg.startswith('condenser.'):
+    if condenser_config_arg.startswith("condenser."):
         condenser_config_arg = condenser_config_arg[10:]
 
     logger.openhands_logger.debug(
-        f'Loading condenser config [{condenser_config_arg}] from {toml_file}'
+        f"Loading condenser config [{condenser_config_arg}] from {toml_file}",
     )
 
     # load the toml file
     try:
-        with open(toml_file, 'r', encoding='utf-8') as toml_contents:
+        with open(toml_file, encoding="utf-8") as toml_contents:
             toml_config = toml.load(toml_contents)
     except FileNotFoundError as e:
-        logger.openhands_logger.error(f'Config file not found: {toml_file}. Error: {e}')
+        logger.openhands_logger.error(f"Config file not found: {toml_file}. Error: {e}")
         return None
     except toml.TomlDecodeError as e:
         logger.openhands_logger.error(
-            f'Cannot parse condenser group [{condenser_config_arg}] from {toml_file}. Exception: {e}'
+            f"Cannot parse condenser group [{condenser_config_arg}] from {toml_file}. Exception: {e}",
         )
         return None
 
     # Check if the condenser section and the specific config exist
     if (
-        'condenser' not in toml_config
-        or condenser_config_arg not in toml_config['condenser']
+        "condenser" not in toml_config
+        or condenser_config_arg not in toml_config["condenser"]
     ):
         logger.openhands_logger.error(
-            f'Condenser config section [condenser.{condenser_config_arg}] not found in {toml_file}'
+            f"Condenser config section [condenser.{condenser_config_arg}] not found in {toml_file}",
         )
         return None
 
-    condenser_data = toml_config['condenser'][
+    condenser_data = toml_config["condenser"][
         condenser_config_arg
     ].copy()  # Use copy to modify
 
     # Determine the type and handle potential LLM dependency
-    condenser_type = condenser_data.get('type')
+    condenser_type = condenser_data.get("type")
     if not condenser_type:
         logger.openhands_logger.error(
-            f'Missing "type" field in [condenser.{condenser_config_arg}] section of {toml_file}'
+            f'Missing "type" field in [condenser.{condenser_config_arg}] section of {toml_file}',
         )
         return None
 
     # Handle LLM config reference if needed, using get_llm_config_arg
     if (
-        condenser_type in ('llm', 'llm_attention', 'structured')
-        and 'llm_config' in condenser_data
-        and isinstance(condenser_data['llm_config'], str)
+        condenser_type in ("llm", "llm_attention", "structured")
+        and "llm_config" in condenser_data
+        and isinstance(condenser_data["llm_config"], str)
     ):
-        llm_config_name = condenser_data['llm_config']
+        llm_config_name = condenser_data["llm_config"]
         logger.openhands_logger.debug(
-            f'Condenser [{condenser_config_arg}] requires LLM config [{llm_config_name}]. Loading it...'
+            f"Condenser [{condenser_config_arg}] requires LLM config [{llm_config_name}]. Loading it...",
         )
         # Use the existing function to load the specific LLM config
         referenced_llm_config = get_llm_config_arg(llm_config_name, toml_file=toml_file)
 
         if referenced_llm_config:
             # Replace the string reference with the actual LLMConfig object
-            condenser_data['llm_config'] = referenced_llm_config
+            condenser_data["llm_config"] = referenced_llm_config
         else:
             # get_llm_config_arg already logs the error if not found
             logger.openhands_logger.error(
-                f"Failed to load required LLM config '{llm_config_name}' for condenser '{condenser_config_arg}'."
+                f"Failed to load required LLM config '{llm_config_name}' for condenser '{condenser_config_arg}'.",
             )
             return None
 
@@ -625,12 +625,12 @@ def get_condenser_config_arg(
     try:
         config = create_condenser_config(condenser_type, condenser_data)
         logger.openhands_logger.info(
-            f'Successfully loaded condenser config [{condenser_config_arg}] from {toml_file}'
+            f"Successfully loaded condenser config [{condenser_config_arg}] from {toml_file}",
         )
         return config
     except (ValidationError, ValueError) as e:
         logger.openhands_logger.error(
-            f'Invalid condenser configuration for [{condenser_config_arg}]: {e}.'
+            f"Invalid condenser configuration for [{condenser_config_arg}]: {e}.",
         )
         return None
 
@@ -638,124 +638,124 @@ def get_condenser_config_arg(
 # Command line arguments
 def get_parser() -> argparse.ArgumentParser:
     """Get the argument parser."""
-    parser = argparse.ArgumentParser(description='Run the agent via CLI')
+    parser = argparse.ArgumentParser(description="Run the agent via CLI")
 
     # Add version argument
     parser.add_argument(
-        '-v', '--version', action='store_true', help='Show version information'
+        "-v", "--version", action="store_true", help="Show version information",
     )
 
     parser.add_argument(
-        '--config-file',
+        "--config-file",
         type=str,
-        default='config.toml',
-        help='Path to the config file (default: config.toml in the current directory)',
+        default="config.toml",
+        help="Path to the config file (default: config.toml in the current directory)",
     )
     parser.add_argument(
-        '-d',
-        '--directory',
+        "-d",
+        "--directory",
         type=str,
-        help='The working directory for the agent',
+        help="The working directory for the agent",
     )
     parser.add_argument(
-        '-t',
-        '--task',
+        "-t",
+        "--task",
         type=str,
-        default='',
-        help='The task for the agent to perform',
+        default="",
+        help="The task for the agent to perform",
     )
     parser.add_argument(
-        '-f',
-        '--file',
+        "-f",
+        "--file",
         type=str,
-        help='Path to a file containing the task. Overrides -t if both are provided.',
+        help="Path to a file containing the task. Overrides -t if both are provided.",
     )
     parser.add_argument(
-        '-c',
-        '--agent-cls',
+        "-c",
+        "--agent-cls",
         default=OH_DEFAULT_AGENT,
         type=str,
-        help='Name of the default agent to use',
+        help="Name of the default agent to use",
     )
     parser.add_argument(
-        '-i',
-        '--max-iterations',
+        "-i",
+        "--max-iterations",
         default=OH_MAX_ITERATIONS,
         type=int,
-        help='The maximum number of iterations to run the agent',
+        help="The maximum number of iterations to run the agent",
     )
     parser.add_argument(
-        '-b',
-        '--max-budget-per-task',
+        "-b",
+        "--max-budget-per-task",
         type=float,
-        help='The maximum budget allowed per task, beyond which the agent will stop.',
+        help="The maximum budget allowed per task, beyond which the agent will stop.",
     )
     # --eval configs are for evaluations only
     parser.add_argument(
-        '--eval-output-dir',
-        default='evaluation/evaluation_outputs/outputs',
+        "--eval-output-dir",
+        default="evaluation/evaluation_outputs/outputs",
         type=str,
-        help='The directory to save evaluation output',
+        help="The directory to save evaluation output",
     )
     parser.add_argument(
-        '--eval-n-limit',
+        "--eval-n-limit",
         default=None,
         type=int,
-        help='The number of instances to evaluate',
+        help="The number of instances to evaluate",
     )
     parser.add_argument(
-        '--eval-num-workers',
+        "--eval-num-workers",
         default=4,
         type=int,
-        help='The number of workers to use for evaluation',
+        help="The number of workers to use for evaluation",
     )
     parser.add_argument(
-        '--eval-note',
+        "--eval-note",
         default=None,
         type=str,
-        help='The note to add to the evaluation directory',
+        help="The note to add to the evaluation directory",
     )
     parser.add_argument(
-        '-l',
-        '--llm-config',
+        "-l",
+        "--llm-config",
         default=None,
         type=str,
         help='Replace default LLM ([llm] section in config.toml) config with the specified LLM config, e.g. "llama3" for [llm.llama3] section in config.toml',
     )
     parser.add_argument(
-        '--agent-config',
+        "--agent-config",
         default=None,
         type=str,
         help='Replace default Agent ([agent] section in config.toml) config with the specified Agent config, e.g. "CodeAct" for [agent.CodeAct] section in config.toml',
     )
     parser.add_argument(
-        '-n',
-        '--name',
-        help='Session name',
+        "-n",
+        "--name",
+        help="Session name",
         type=str,
-        default='',
+        default="",
     )
     parser.add_argument(
-        '--eval-ids',
+        "--eval-ids",
         default=None,
         type=str,
-        help='The comma-separated list (in quotes) of IDs of the instances to evaluate',
+        help="The comma-separated list (in quotes) of IDs of the instances to evaluate",
     )
     parser.add_argument(
-        '--no-auto-continue',
-        help='Disable auto-continue responses in headless mode (i.e. headless will read from stdin instead of auto-continuing)',
-        action='store_true',
+        "--no-auto-continue",
+        help="Disable auto-continue responses in headless mode (i.e. headless will read from stdin instead of auto-continuing)",
+        action="store_true",
         default=False,
     )
     parser.add_argument(
-        '--selected-repo',
-        help='GitHub repository to clone (format: owner/repo)',
+        "--selected-repo",
+        help="GitHub repository to clone (format: owner/repo)",
         type=str,
         default=None,
     )
     parser.add_argument(
-        '--override-cli-mode',
-        help='Override the default settings for CLI mode',
+        "--override-cli-mode",
+        help="Override the default settings for CLI mode",
         type=bool,
         default=False,
     )
@@ -768,7 +768,7 @@ def parse_arguments() -> argparse.Namespace:
     args = parser.parse_args()
 
     if args.version:
-        print(f'OpenHands version: {__version__}')
+        print(f"OpenHands version: {__version__}")
         sys.exit(0)
 
     return args
@@ -789,16 +789,16 @@ def register_custom_agents(config: OpenHandsConfig) -> None:
                 agent_cls = get_impl(Agent, agent_config.classpath)
                 Agent.register(agent_name, agent_cls)
                 logger.openhands_logger.info(
-                    f"Registered custom agent '{agent_name}' from {agent_config.classpath}"
+                    f"Registered custom agent '{agent_name}' from {agent_config.classpath}",
                 )
             except Exception as e:
                 logger.openhands_logger.error(
-                    f"Failed to register agent '{agent_name}': {e}"
+                    f"Failed to register agent '{agent_name}': {e}",
                 )
 
 
 def load_openhands_config(
-    set_logging_levels: bool = True, config_file: str = 'config.toml'
+    set_logging_levels: bool = True, config_file: str = "config.toml",
 ) -> OpenHandsConfig:
     """Load the configuration from the specified config file and environment variables.
 
@@ -833,7 +833,7 @@ def setup_config_from_args(args: argparse.Namespace) -> OpenHandsConfig:
         else:
             llm_config = config.llms[args.llm_config]
         if llm_config is None:
-            raise ValueError(f'Invalid toml file, cannot read {args.llm_config}')
+            raise ValueError(f"Invalid toml file, cannot read {args.llm_config}")
         config.set_llm_config(llm_config)
 
     # Override default agent if provided
