@@ -6,16 +6,15 @@
 class AgentError(Exception):
     """Base class for all agent exceptions."""
 
-    pass
 
 
 class AgentNoInstructionError(AgentError):
-    def __init__(self, message: str = 'Instruction must be provided') -> None:
+    def __init__(self, message: str = "Instruction must be provided") -> None:
         super().__init__(message)
 
 
 class AgentEventTypeError(AgentError):
-    def __init__(self, message: str = 'Event must be a dictionary') -> None:
+    def __init__(self, message: str = "Event must be a dictionary") -> None:
         super().__init__(message)
 
 
@@ -24,7 +23,7 @@ class AgentAlreadyRegisteredError(AgentError):
         if name is not None:
             message = f"Agent class already registered under '{name}'"
         else:
-            message = 'Agent class already registered'
+            message = "Agent class already registered"
         super().__init__(message)
 
 
@@ -33,12 +32,12 @@ class AgentNotRegisteredError(AgentError):
         if name is not None:
             message = f"No agent class registered under '{name}'"
         else:
-            message = 'No agent class registered'
+            message = "No agent class registered"
         super().__init__(message)
 
 
 class AgentStuckInLoopError(AgentError):
-    def __init__(self, message: str = 'Agent got stuck in a loop') -> None:
+    def __init__(self, message: str = "Agent got stuck in a loop") -> None:
         super().__init__(message)
 
 
@@ -50,9 +49,9 @@ class AgentStuckInLoopError(AgentError):
 class TaskInvalidStateError(Exception):
     def __init__(self, state: str | None = None) -> None:
         if state is not None:
-            message = f'Invalid state {state}'
+            message = f"Invalid state {state}"
         else:
-            message = 'Invalid state'
+            message = "Invalid state"
         super().__init__(message)
 
 
@@ -64,7 +63,7 @@ class TaskInvalidStateError(Exception):
 # This exception gets sent back to the LLM
 # It might be malformed JSON
 class LLMMalformedActionError(Exception):
-    def __init__(self, message: str = 'Malformed response') -> None:
+    def __init__(self, message: str = "Malformed response") -> None:
         self.message = message
         super().__init__(message)
 
@@ -75,7 +74,7 @@ class LLMMalformedActionError(Exception):
 # This exception gets sent back to the LLM
 # For some reason, the agent did not return an action
 class LLMNoActionError(Exception):
-    def __init__(self, message: str = 'Agent must return an action') -> None:
+    def __init__(self, message: str = "Agent must return an action") -> None:
         super().__init__(message)
 
 
@@ -83,7 +82,7 @@ class LLMNoActionError(Exception):
 # The LLM output did not include an action, or the action was not the expected type
 class LLMResponseError(Exception):
     def __init__(
-        self, message: str = 'Failed to retrieve action from LLM response'
+        self, message: str = "Failed to retrieve action from LLM response",
     ) -> None:
         super().__init__(message)
 
@@ -93,27 +92,27 @@ class LLMResponseError(Exception):
 class LLMNoResponseError(Exception):
     def __init__(
         self,
-        message: str = 'LLM did not return a response. This is only seen in Gemini models so far.',
+        message: str = "LLM did not return a response. This is only seen in Gemini models so far.",
     ) -> None:
         super().__init__(message)
 
 
 class UserCancelledError(Exception):
-    def __init__(self, message: str = 'User cancelled the request') -> None:
+    def __init__(self, message: str = "User cancelled the request") -> None:
         super().__init__(message)
 
 
 class OperationCancelled(Exception):
     """Exception raised when an operation is cancelled (e.g. by a keyboard interrupt)."""
 
-    def __init__(self, message: str = 'Operation was cancelled') -> None:
+    def __init__(self, message: str = "Operation was cancelled") -> None:
         super().__init__(message)
 
 
 class LLMContextWindowExceedError(RuntimeError):
     def __init__(
         self,
-        message: str = 'Conversation history longer than LLM context window limit. Consider turning on enable_history_truncation config to avoid this error',
+        message: str = "Conversation history longer than LLM context window limit. Consider turning on enable_history_truncation config to avoid this error",
     ) -> None:
         super().__init__(message)
 
@@ -158,43 +157,36 @@ class FunctionCallNotExistsError(Exception):
 class AgentRuntimeError(Exception):
     """Base class for all agent runtime exceptions."""
 
-    pass
 
 
 class AgentRuntimeBuildError(AgentRuntimeError):
     """Exception raised when an agent runtime build operation fails."""
 
-    pass
 
 
 class AgentRuntimeTimeoutError(AgentRuntimeError):
     """Exception raised when an agent runtime operation times out."""
 
-    pass
 
 
 class AgentRuntimeUnavailableError(AgentRuntimeError):
     """Exception raised when an agent runtime is unavailable."""
 
-    pass
 
 
 class AgentRuntimeNotReadyError(AgentRuntimeUnavailableError):
     """Exception raised when an agent runtime is not ready."""
 
-    pass
 
 
 class AgentRuntimeDisconnectedError(AgentRuntimeUnavailableError):
     """Exception raised when an agent runtime is disconnected."""
 
-    pass
 
 
 class AgentRuntimeNotFoundError(AgentRuntimeUnavailableError):
     """Exception raised when an agent runtime is not found."""
 
-    pass
 
 
 # ============================================
@@ -204,7 +196,7 @@ class AgentRuntimeNotFoundError(AgentRuntimeUnavailableError):
 
 class BrowserInitException(Exception):
     def __init__(
-        self, message: str = 'Failed to initialize browser environment'
+        self, message: str = "Failed to initialize browser environment",
     ) -> None:
         super().__init__(message)
 
@@ -212,7 +204,7 @@ class BrowserInitException(Exception):
 class BrowserUnavailableException(Exception):
     def __init__(
         self,
-        message: str = 'Browser environment is not available, please check if has been initialized',
+        message: str = "Browser environment is not available, please check if has been initialized",
     ) -> None:
         super().__init__(message)
 
@@ -225,11 +217,10 @@ class BrowserUnavailableException(Exception):
 class MicroagentError(Exception):
     """Base exception for all microagent errors."""
 
-    pass
 
 
 class MicroagentValidationError(MicroagentError):
     """Raised when there's a validation error in microagent metadata."""
 
-    def __init__(self, message: str = 'Microagent validation failed') -> None:
+    def __init__(self, message: str = "Microagent validation failed") -> None:
         super().__init__(message)
