@@ -20,7 +20,7 @@ async def run_agent_until_done(
     """
 
     def status_callback(msg_type: str, msg_id: str, msg: str) -> None:
-        if msg_type == 'error':
+        if msg_type == "error":
             logger.error(msg)
             if controller:
                 controller.state.last_error = msg
@@ -28,13 +28,13 @@ async def run_agent_until_done(
         else:
             logger.info(msg)
 
-    if hasattr(runtime, 'status_callback') and runtime.status_callback:
+    if hasattr(runtime, "status_callback") and runtime.status_callback:
         raise ValueError(
-            'Runtime status_callback was set, but run_agent_until_done will override it'
+            "Runtime status_callback was set, but run_agent_until_done will override it",
         )
-    if hasattr(controller, 'status_callback') and controller.status_callback:
+    if hasattr(controller, "status_callback") and controller.status_callback:
         raise ValueError(
-            'Controller status_callback was set, but run_agent_until_done will override it'
+            "Controller status_callback was set, but run_agent_until_done will override it",
         )
 
     runtime.status_callback = status_callback
