@@ -3,7 +3,7 @@ from openhands.core.schema.agent import AgentState
 from openhands.core.schema.observation import ObservationType
 
 __all__ = [
-    'ActionType',
-    'ObservationType',
-    'AgentState',
+    "ActionType",
+    "AgentState",
+    "ObservationType",
 ]
