@@ -2,92 +2,92 @@ from enum import Enum
 
 
 class ActionType(str, Enum):
-    MESSAGE = 'message'
+    MESSAGE = "message"
     """Represents a message.
     """
 
-    SYSTEM = 'system'
+    SYSTEM = "system"
     """Represents a system message.
     """
 
-    START = 'start'
+    START = "start"
     """Starts a new development task OR send chat from the user. Only sent by the client.
     """
 
-    READ = 'read'
+    READ = "read"
     """Reads the content of a file.
     """
 
-    WRITE = 'write'
+    WRITE = "write"
     """Writes the content to a file.
     """
 
-    EDIT = 'edit'
+    EDIT = "edit"
     """Edits a file by providing a draft.
     """
 
-    RUN = 'run'
+    RUN = "run"
     """Runs a command.
     """
 
-    RUN_IPYTHON = 'run_ipython'
+    RUN_IPYTHON = "run_ipython"
     """Runs a IPython cell.
     """
 
-    BROWSE = 'browse'
+    BROWSE = "browse"
     """Opens a web page.
     """
 
-    BROWSE_INTERACTIVE = 'browse_interactive'
+    BROWSE_INTERACTIVE = "browse_interactive"
     """Interact with the browser instance.
     """
 
-    MCP = 'call_tool_mcp'
+    MCP = "call_tool_mcp"
     """Interact with the MCP server.
     """
 
-    DELEGATE = 'delegate'
+    DELEGATE = "delegate"
     """Delegates a task to another agent.
     """
 
-    THINK = 'think'
+    THINK = "think"
     """Logs a thought.
     """
 
-    FINISH = 'finish'
+    FINISH = "finish"
     """If you're absolutely certain that you've completed your task and have tested your work,
     use the finish action to stop working.
     """
 
-    REJECT = 'reject'
+    REJECT = "reject"
     """If you're absolutely certain that you cannot complete the task with given requirements,
     use the reject action to stop working.
     """
 
-    NULL = 'null'
+    NULL = "null"
 
-    PAUSE = 'pause'
+    PAUSE = "pause"
     """Pauses the task.
     """
 
-    RESUME = 'resume'
+    RESUME = "resume"
     """Resumes the task.
     """
 
-    STOP = 'stop'
+    STOP = "stop"
     """Stops the task. Must send a start action to restart a new task.
     """
 
-    CHANGE_AGENT_STATE = 'change_agent_state'
+    CHANGE_AGENT_STATE = "change_agent_state"
 
-    PUSH = 'push'
+    PUSH = "push"
     """Push a branch to github."""
 
-    SEND_PR = 'send_pr'
+    SEND_PR = "send_pr"
     """Send a PR to github."""
 
-    RECALL = 'recall'
+    RECALL = "recall"
     """Retrieves content from a user workspace, microagent, or other source."""
 
-    CONDENSATION = 'condensation'
+    CONDENSATION = "condensation"
     """Condenses a list of events into a summary."""
