@@ -2,50 +2,50 @@ from enum import Enum
 
 
 class AgentState(str, Enum):
-    LOADING = 'loading'
+    LOADING = "loading"
     """The agent is loading.
     """
 
-    RUNNING = 'running'
+    RUNNING = "running"
     """The agent is running.
     """
 
-    AWAITING_USER_INPUT = 'awaiting_user_input'
+    AWAITING_USER_INPUT = "awaiting_user_input"
     """The agent is awaiting user input.
     """
 
-    PAUSED = 'paused'
+    PAUSED = "paused"
     """The agent is paused.
     """
 
-    STOPPED = 'stopped'
+    STOPPED = "stopped"
     """The agent is stopped.
     """
 
-    FINISHED = 'finished'
+    FINISHED = "finished"
     """The agent is finished with the current task.
     """
 
-    REJECTED = 'rejected'
+    REJECTED = "rejected"
     """The agent rejects the task.
     """
 
-    ERROR = 'error'
+    ERROR = "error"
     """An error occurred during the task.
     """
 
-    AWAITING_USER_CONFIRMATION = 'awaiting_user_confirmation'
+    AWAITING_USER_CONFIRMATION = "awaiting_user_confirmation"
     """The agent is awaiting user confirmation.
     """
 
-    USER_CONFIRMED = 'user_confirmed'
+    USER_CONFIRMED = "user_confirmed"
     """The user confirmed the agent's action.
     """
 
-    USER_REJECTED = 'user_rejected'
+    USER_REJECTED = "user_rejected"
     """The user rejected the agent's action.
     """
 
-    RATE_LIMITED = 'rate_limited'
+    RATE_LIMITED = "rate_limited"
     """The agent is rate limited.
     """
