@@ -2,9 +2,9 @@ from openhands.events.event import Event, EventSource, RecallType
 from openhands.events.stream import EventStream, EventStreamSubscriber
 
 __all__ = [
-    'Event',
-    'EventSource',
-    'EventStream',
-    'EventStreamSubscriber',
-    'RecallType',
+    "Event",
+    "EventSource",
+    "EventStream",
+    "EventStreamSubscriber",
+    "RecallType",
 ]
