@@ -19,23 +19,23 @@ from openhands.events.action.mcp import MCPAction
 from openhands.events.action.message import MessageAction, SystemMessageAction
 
 __all__ = [
-    'Action',
-    'NullAction',
-    'CmdRunAction',
-    'BrowseURLAction',
-    'BrowseInteractiveAction',
-    'FileReadAction',
-    'FileWriteAction',
-    'FileEditAction',
-    'AgentFinishAction',
-    'AgentRejectAction',
-    'AgentDelegateAction',
-    'ChangeAgentStateAction',
-    'IPythonRunCellAction',
-    'MessageAction',
-    'SystemMessageAction',
-    'ActionConfirmationStatus',
-    'AgentThinkAction',
-    'RecallAction',
-    'MCPAction',
+    "Action",
+    "ActionConfirmationStatus",
+    "AgentDelegateAction",
+    "AgentFinishAction",
+    "AgentRejectAction",
+    "AgentThinkAction",
+    "BrowseInteractiveAction",
+    "BrowseURLAction",
+    "ChangeAgentStateAction",
+    "CmdRunAction",
+    "FileEditAction",
+    "FileReadAction",
+    "FileWriteAction",
+    "IPythonRunCellAction",
+    "MCPAction",
+    "MessageAction",
+    "NullAction",
+    "RecallAction",
+    "SystemMessageAction",
 ]
