@@ -12,18 +12,18 @@ class ChangeAgentStateAction(Action):
     """Fake action, just to notify the client that a task state has changed."""
 
     agent_state: str
-    thought: str = ''
+    thought: str = ""
     action: str = ActionType.CHANGE_AGENT_STATE
 
     @property
     def message(self) -> str:
-        return f'Agent state changed to {self.agent_state}'
+        return f"Agent state changed to {self.agent_state}"
 
 
 class AgentFinishTaskCompleted(Enum):
-    FALSE = 'false'
-    PARTIAL = 'partial'
-    TRUE = 'true'
+    FALSE = "false"
+    PARTIAL = "partial"
+    TRUE = "true"
 
 
 @dataclass
@@ -38,15 +38,15 @@ class AgentFinishAction(Action):
         action (str): The action type, namely ActionType.FINISH.
     """
 
-    final_thought: str = ''
+    final_thought: str = ""
     task_completed: AgentFinishTaskCompleted | None = None
     outputs: dict[str, Any] = field(default_factory=dict)
-    thought: str = ''
+    thought: str = ""
     action: str = ActionType.FINISH
 
     @property
     def message(self) -> str:
-        if self.thought != '':
+        if self.thought != "":
             return self.thought
         return "All done! What's next on the agenda?"
 
@@ -60,25 +60,25 @@ class AgentThinkAction(Action):
         action (str): The action type, namely ActionType.THINK.
     """
 
-    thought: str = ''
+    thought: str = ""
     action: str = ActionType.THINK
 
     @property
     def message(self) -> str:
-        return f'I am thinking...: {self.thought}'
+        return f"I am thinking...: {self.thought}"
 
 
 @dataclass
 class AgentRejectAction(Action):
     outputs: dict = field(default_factory=dict)
-    thought: str = ''
+    thought: str = ""
     action: str = ActionType.REJECT
 
     @property
     def message(self) -> str:
-        msg: str = 'Task is rejected by the agent.'
-        if 'reason' in self.outputs:
-            msg += ' Reason: ' + self.outputs['reason']
+        msg: str = "Task is rejected by the agent."
+        if "reason" in self.outputs:
+            msg += " Reason: " + self.outputs["reason"]
         return msg
 
 
@@ -86,7 +86,7 @@ class AgentRejectAction(Action):
 class AgentDelegateAction(Action):
     agent: str
     inputs: dict
-    thought: str = ''
+    thought: str = ""
     action: str = ActionType.DELEGATE
 
     @property
@@ -99,17 +99,17 @@ class RecallAction(Action):
     """This action is used for retrieving content, e.g., from the global directory or user workspace."""
 
     recall_type: RecallType
-    query: str = ''
-    thought: str = ''
+    query: str = ""
+    thought: str = ""
     action: str = ActionType.RECALL
 
     @property
     def message(self) -> str:
-        return f'Retrieving content for: {self.query[:50]}'
+        return f"Retrieving content for: {self.query[:50]}"
 
     def __str__(self) -> str:
-        ret = '**RecallAction**\n'
-        ret += f'QUERY: {self.query[:50]}'
+        ret = "**RecallAction**\n"
+        ret += f"QUERY: {self.query[:50]}"
         return ret
 
 
@@ -168,7 +168,7 @@ class CondensationAction(Action):
 
     def __post_init__(self):
         if not self._validate_field_polymorphism():
-            raise ValueError('Invalid configuration of the optional fields.')
+            raise ValueError("Invalid configuration of the optional fields.")
 
     @property
     def forgotten(self) -> list[int]:
@@ -178,7 +178,7 @@ class CondensationAction(Action):
         # can't make the dataclass immutable so we need to check it again here
         # to make sure the configuration is still valid.
         if not self._validate_field_polymorphism():
-            raise ValueError('Invalid configuration of the optional fields.')
+            raise ValueError("Invalid configuration of the optional fields.")
 
         if self.forgotten_event_ids is not None:
             return self.forgotten_event_ids
@@ -187,11 +187,11 @@ class CondensationAction(Action):
         assert self.forgotten_events_start_id is not None
         assert self.forgotten_events_end_id is not None
         return list(
-            range(self.forgotten_events_start_id, self.forgotten_events_end_id + 1)
+            range(self.forgotten_events_start_id, self.forgotten_events_end_id + 1),
         )
 
     @property
     def message(self) -> str:
         if self.summary:
-            return f'Summary: {self.summary}'
-        return f'Condenser is dropping the events: {self.forgotten}.'
+            return f"Summary: {self.summary}"
+        return f"Condenser is dropping the events: {self.forgotten}."
