@@ -15,7 +15,7 @@ class CmdRunAction(Action):
         str  # When `command` is empty, it will be used to print the current tmux window
     )
     is_input: bool = False  # if True, the command is an input to the running process
-    thought: str = ''
+    thought: str = ""
     blocking: bool = False  # if True, the command will be run in a blocking manner, but a timeout must be set through _set_hard_timeout
     is_static: bool = False  # if True, runs the command in a separate process
     cwd: str | None = None  # current working directory, only used if is_static is True
@@ -27,20 +27,20 @@ class CmdRunAction(Action):
 
     @property
     def message(self) -> str:
-        return f'Running command: {self.command}'
+        return f"Running command: {self.command}"
 
     def __str__(self) -> str:
-        ret = f'**CmdRunAction (source={self.source}, is_input={self.is_input})**\n'
+        ret = f"**CmdRunAction (source={self.source}, is_input={self.is_input})**\n"
         if self.thought:
-            ret += f'THOUGHT: {self.thought}\n'
-        ret += f'COMMAND:\n{self.command}'
+            ret += f"THOUGHT: {self.thought}\n"
+        ret += f"COMMAND:\n{self.command}"
         return ret
 
 
 @dataclass
 class IPythonRunCellAction(Action):
     code: str
-    thought: str = ''
+    thought: str = ""
     include_extra: bool = (
         True  # whether to include CWD & Python interpreter in the output
     )
@@ -48,15 +48,15 @@ class IPythonRunCellAction(Action):
     runnable: ClassVar[bool] = True
     confirmation_state: ActionConfirmationStatus = ActionConfirmationStatus.CONFIRMED
     security_risk: ActionSecurityRisk | None = None
-    kernel_init_code: str = ''  # code to run in the kernel (if the kernel is restarted)
+    kernel_init_code: str = ""  # code to run in the kernel (if the kernel is restarted)
 
     def __str__(self) -> str:
-        ret = '**IPythonRunCellAction**\n'
+        ret = "**IPythonRunCellAction**\n"
         if self.thought:
-            ret += f'THOUGHT: {self.thought}\n'
-        ret += f'CODE:\n{self.code}'
+            ret += f"THOUGHT: {self.thought}\n"
+        ret += f"CODE:\n{self.code}"
         return ret
 
     @property
     def message(self) -> str:
-        return f'Running Python code interactively: {self.code}'
+        return f"Running Python code interactively: {self.code}"
