@@ -27,23 +27,23 @@ from openhands.events.observation.reject import UserRejectObservation
 from openhands.events.observation.success import SuccessObservation
 
 __all__ = [
-    'Observation',
-    'NullObservation',
-    'AgentThinkObservation',
-    'CmdOutputObservation',
-    'CmdOutputMetadata',
-    'IPythonRunCellObservation',
-    'BrowserOutputObservation',
-    'FileReadObservation',
-    'FileWriteObservation',
-    'FileEditObservation',
-    'ErrorObservation',
-    'AgentStateChangedObservation',
-    'AgentDelegateObservation',
-    'SuccessObservation',
-    'UserRejectObservation',
-    'AgentCondensationObservation',
-    'RecallObservation',
-    'RecallType',
-    'MCPObservation',
+    "AgentCondensationObservation",
+    "AgentDelegateObservation",
+    "AgentStateChangedObservation",
+    "AgentThinkObservation",
+    "BrowserOutputObservation",
+    "CmdOutputMetadata",
+    "CmdOutputObservation",
+    "ErrorObservation",
+    "FileEditObservation",
+    "FileReadObservation",
+    "FileWriteObservation",
+    "IPythonRunCellObservation",
+    "MCPObservation",
+    "NullObservation",
+    "Observation",
+    "RecallObservation",
+    "RecallType",
+    "SuccessObservation",
+    "UserRejectObservation",
 ]
