@@ -10,12 +10,12 @@ class AgentStateChangedObservation(Observation):
     """This data class represents the result from delegating to another agent"""
 
     agent_state: str
-    reason: str = ''
+    reason: str = ""
     observation: str = ObservationType.AGENT_STATE_CHANGED
 
     @property
     def message(self) -> str:
-        return ''
+        return ""
 
 
 @dataclass
@@ -68,14 +68,14 @@ class RecallObservation(Observation):
     observation: str = ObservationType.RECALL
 
     # workspace context
-    repo_name: str = ''
-    repo_directory: str = ''
-    repo_instructions: str = ''
+    repo_name: str = ""
+    repo_directory: str = ""
+    repo_instructions: str = ""
     runtime_hosts: dict[str, int] = field(default_factory=dict)
-    additional_agent_instructions: str = ''
-    date: str = ''
+    additional_agent_instructions: str = ""
+    date: str = ""
     custom_secrets_descriptions: dict[str, str] = field(default_factory=dict)
-    conversation_instructions: str = ''
+    conversation_instructions: str = ""
 
     # knowledge
     microagent_knowledge: list[MicroagentKnowledge] = field(default_factory=list)
@@ -100,9 +100,9 @@ class RecallObservation(Observation):
     @property
     def message(self) -> str:
         return (
-            'Added workspace context'
+            "Added workspace context"
             if self.recall_type == RecallType.WORKSPACE_CONTEXT
-            else 'Added microagent knowledge'
+            else "Added microagent knowledge"
         )
 
     def __str__(self) -> str:
@@ -111,27 +111,27 @@ class RecallObservation(Observation):
         if self.recall_type == RecallType.WORKSPACE_CONTEXT:
             fields.extend(
                 [
-                    f'recall_type={self.recall_type}',
-                    f'repo_name={self.repo_name}',
-                    f'repo_instructions={self.repo_instructions[:20]}...',
-                    f'runtime_hosts={self.runtime_hosts}',
-                    f'additional_agent_instructions={self.additional_agent_instructions[:20]}...',
-                    f'date={self.date}'
-                    f'custom_secrets_descriptions={self.custom_secrets_descriptions}',
-                    f'conversation_instructions={self.conversation_instructions[0:20]}...',
-                ]
+                    f"recall_type={self.recall_type}",
+                    f"repo_name={self.repo_name}",
+                    f"repo_instructions={self.repo_instructions[:20]}...",
+                    f"runtime_hosts={self.runtime_hosts}",
+                    f"additional_agent_instructions={self.additional_agent_instructions[:20]}...",
+                    f"date={self.date}"
+                    f"custom_secrets_descriptions={self.custom_secrets_descriptions}",
+                    f"conversation_instructions={self.conversation_instructions[0:20]}...",
+                ],
             )
         else:
             fields.extend(
                 [
-                    f'recall_type={self.recall_type}',
-                ]
+                    f"recall_type={self.recall_type}",
+                ],
             )
         if self.microagent_knowledge:
             fields.extend(
                 [
                     f'microagent_knowledge={", ".join([m.name for m in self.microagent_knowledge])}',
-                ]
+                ],
             )
 
         return f'**RecallObservation**\n{", ".join(fields)}'
