@@ -11,9 +11,9 @@ class BrowserOutputObservation(Observation):
 
     url: str
     trigger_by_action: str
-    screenshot: str = field(repr=False, default='')  # don't show in repr
+    screenshot: str = field(repr=False, default="")  # don't show in repr
     screenshot_path: str | None = field(default=None)  # path to saved screenshot file
-    set_of_marks: str = field(default='', repr=False)  # don't show in repr
+    set_of_marks: str = field(default="", repr=False)  # don't show in repr
     error: bool = False
     observation: str = ObservationType.BROWSE
     goal_image_urls: list[str] = field(default_factory=list)
@@ -21,35 +21,35 @@ class BrowserOutputObservation(Observation):
     open_pages_urls: list[str] = field(default_factory=list)
     active_page_index: int = -1
     dom_object: dict[str, Any] = field(
-        default_factory=dict, repr=False
+        default_factory=dict, repr=False,
     )  # don't show in repr
     axtree_object: dict[str, Any] = field(
-        default_factory=dict, repr=False
+        default_factory=dict, repr=False,
     )  # don't show in repr
     extra_element_properties: dict[str, Any] = field(
-        default_factory=dict, repr=False
+        default_factory=dict, repr=False,
     )  # don't show in repr
-    last_browser_action: str = ''
-    last_browser_action_error: str = ''
-    focused_element_bid: str = ''
+    last_browser_action: str = ""
+    last_browser_action_error: str = ""
+    focused_element_bid: str = ""
 
     @property
     def message(self) -> str:
-        return 'Visited ' + self.url
+        return "Visited " + self.url
 
     def __str__(self) -> str:
         ret = (
-            '**BrowserOutputObservation**\n'
-            f'URL: {self.url}\n'
-            f'Error: {self.error}\n'
-            f'Open pages: {self.open_pages_urls}\n'
-            f'Active page index: {self.active_page_index}\n'
-            f'Last browser action: {self.last_browser_action}\n'
-            f'Last browser action error: {self.last_browser_action_error}\n'
-            f'Focused element bid: {self.focused_element_bid}\n'
+            "**BrowserOutputObservation**\n"
+            f"URL: {self.url}\n"
+            f"Error: {self.error}\n"
+            f"Open pages: {self.open_pages_urls}\n"
+            f"Active page index: {self.active_page_index}\n"
+            f"Last browser action: {self.last_browser_action}\n"
+            f"Last browser action error: {self.last_browser_action_error}\n"
+            f"Focused element bid: {self.focused_element_bid}\n"
         )
         if self.screenshot_path:
-            ret += f'Screenshot saved to: {self.screenshot_path}\n'
-        ret += '--- Agent Observation ---\n'
+            ret += f"Screenshot saved to: {self.screenshot_path}\n"
+        ret += "--- Agent Observation ---\n"
         ret += self.content
         return ret
