@@ -10,10 +10,10 @@ from openhands.core.logger import openhands_logger as logger
 from openhands.core.schema import ObservationType
 from openhands.events.observation.observation import Observation
 
-CMD_OUTPUT_PS1_BEGIN = '\n###PS1JSON###\n'
-CMD_OUTPUT_PS1_END = '\n###PS1END###'
+CMD_OUTPUT_PS1_BEGIN = "\n###PS1JSON###\n"
+CMD_OUTPUT_PS1_END = "\n###PS1END###"
 CMD_OUTPUT_METADATA_PS1_REGEX = re.compile(
-    f'^{CMD_OUTPUT_PS1_BEGIN.strip()}(.*?){CMD_OUTPUT_PS1_END.strip()}',
+    f"^{CMD_OUTPUT_PS1_BEGIN.strip()}(.*?){CMD_OUTPUT_PS1_END.strip()}",
     re.DOTALL | re.MULTILINE,
 )
 
@@ -27,8 +27,8 @@ class CmdOutputMetadata(BaseModel):
     hostname: str | None = None
     working_dir: str | None = None
     py_interpreter_path: str | None = None
-    prefix: str = ''  # Prefix to add to command output
-    suffix: str = ''  # Suffix to add to command output
+    prefix: str = ""  # Prefix to add to command output
+    suffix: str = ""  # Suffix to add to command output
 
     @classmethod
     def to_ps1_prompt(cls) -> str:
@@ -36,19 +36,19 @@ class CmdOutputMetadata(BaseModel):
         prompt = CMD_OUTPUT_PS1_BEGIN
         json_str = json.dumps(
             {
-                'pid': '$!',
-                'exit_code': '$?',
-                'username': r'\u',
-                'hostname': r'\h',
-                'working_dir': r'$(pwd)',
-                'py_interpreter_path': r'$(which python 2>/dev/null || echo "")',
+                "pid": "$!",
+                "exit_code": "$?",
+                "username": r"\u",
+                "hostname": r"\h",
+                "working_dir": r"$(pwd)",
+                "py_interpreter_path": r'$(which python 2>/dev/null || echo "")',
             },
             indent=2,
         )
         # Make sure we escape double quotes in the JSON string
         # So that PS1 will keep them as part of the output
         prompt += json_str.replace('"', r'\"')
-        prompt += CMD_OUTPUT_PS1_END + '\n'  # Ensure there's a newline at the end
+        prompt += CMD_OUTPUT_PS1_END + "\n"  # Ensure there's a newline at the end
         return prompt
 
     @classmethod
@@ -60,8 +60,8 @@ class CmdOutputMetadata(BaseModel):
                 matches.append(match)
             except json.JSONDecodeError:
                 logger.warning(
-                    f'Failed to parse PS1 metadata: {match.group(1)}. Skipping.'
-                    + traceback.format_exc()
+                    f"Failed to parse PS1 metadata: {match.group(1)}. Skipping."
+                    + traceback.format_exc(),
                 )
                 continue  # Skip if not valid JSON
         return matches
@@ -73,19 +73,19 @@ class CmdOutputMetadata(BaseModel):
         # Create a copy of metadata to avoid modifying the original
         processed = metadata.copy()
         # Convert numeric fields
-        if 'pid' in metadata:
+        if "pid" in metadata:
             try:
-                processed['pid'] = int(float(str(metadata['pid'])))
+                processed["pid"] = int(float(str(metadata["pid"])))
             except (ValueError, TypeError):
-                processed['pid'] = -1
-        if 'exit_code' in metadata:
+                processed["pid"] = -1
+        if "exit_code" in metadata:
             try:
-                processed['exit_code'] = int(float(str(metadata['exit_code'])))
+                processed["exit_code"] = int(float(str(metadata["exit_code"])))
             except (ValueError, TypeError):
                 logger.warning(
-                    f'Failed to parse exit code: {metadata["exit_code"]}. Setting to -1.'
+                    f'Failed to parse exit code: {metadata["exit_code"]}. Setting to -1.',
                 )
-                processed['exit_code'] = -1
+                processed["exit_code"] = -1
         return cls(**processed)
 
 
@@ -119,10 +119,10 @@ class CmdOutputObservation(Observation):
             self.metadata = metadata or CmdOutputMetadata()
 
         # Handle legacy attribute
-        if 'exit_code' in kwargs:
-            self.metadata.exit_code = kwargs['exit_code']
-        if 'command_id' in kwargs:
-            self.metadata.pid = kwargs['command_id']
+        if "exit_code" in kwargs:
+            self.metadata.exit_code = kwargs["exit_code"]
+        if "command_id" in kwargs:
+            self.metadata.pid = kwargs["command_id"]
 
     @property
     def command_id(self) -> int:
@@ -138,7 +138,7 @@ class CmdOutputObservation(Observation):
 
     @property
     def message(self) -> str:
-        return f'Command `{self.command}` executed with exit code {self.exit_code}.'
+        return f"Command `{self.command}` executed with exit code {self.exit_code}."
 
     @property
     def success(self) -> bool:
@@ -146,21 +146,21 @@ class CmdOutputObservation(Observation):
 
     def __str__(self) -> str:
         return (
-            f'**CmdOutputObservation (source={self.source}, exit code={self.exit_code}, '
-            f'metadata={json.dumps(self.metadata.model_dump(), indent=2)})**\n'
-            '--BEGIN AGENT OBSERVATION--\n'
-            f'{self.to_agent_observation()}\n'
-            '--END AGENT OBSERVATION--'
+            f"**CmdOutputObservation (source={self.source}, exit code={self.exit_code}, "
+            f"metadata={json.dumps(self.metadata.model_dump(), indent=2)})**\n"
+            "--BEGIN AGENT OBSERVATION--\n"
+            f"{self.to_agent_observation()}\n"
+            "--END AGENT OBSERVATION--"
         )
 
     def to_agent_observation(self) -> str:
-        ret = f'{self.metadata.prefix}{self.content}{self.metadata.suffix}'
+        ret = f"{self.metadata.prefix}{self.content}{self.metadata.suffix}"
         if self.metadata.working_dir:
-            ret += f'\n[Current working directory: {self.metadata.working_dir}]'
+            ret += f"\n[Current working directory: {self.metadata.working_dir}]"
         if self.metadata.py_interpreter_path:
-            ret += f'\n[Python interpreter: {self.metadata.py_interpreter_path}]'
+            ret += f"\n[Python interpreter: {self.metadata.py_interpreter_path}]"
         if self.metadata.exit_code != -1:
-            ret += f'\n[Command finished with exit code {self.metadata.exit_code}]'
+            ret += f"\n[Command finished with exit code {self.metadata.exit_code}]"
         return ret
 
 
@@ -178,14 +178,14 @@ class IPythonRunCellObservation(Observation):
 
     @property
     def message(self) -> str:
-        return 'Code executed in IPython cell.'
+        return "Code executed in IPython cell."
 
     @property
     def success(self) -> bool:
         return True  # IPython cells are always considered successful
 
     def __str__(self) -> str:
-        result = f'**IPythonRunCellObservation**\n{self.content}'
+        result = f"**IPythonRunCellObservation**\n{self.content}"
         if self.image_urls:
-            result += f'\nImages: {len(self.image_urls)}'
+            result += f"\nImages: {len(self.image_urls)}"
         return result
